% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{bcs_planner}
\alias{bcs_planner}
\title{The default way to make multiprocess plans for the bcs-cycle1 server}
\usage{
bcs_planner(login_node, core_function = future::availableCores,
  backoff_threshold = 0.7)
}
\arguments{
\item{login_node}{a string for the gateway login, e.g., "zachburchill@cycle1.cs.rochester.edu"}

\item{core_function}{the function determining the number of cores to use on the server. By default, it is the default of \code{multiprocess}. You can specify a custom function or just set a number yourself}

\item{backoff_threshold}{the threshold of percentage of available memory below which the connection is cancelled}
}
\description{
If one wants to run jobs on the bcs-cycle1 server.
}
