% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{\%beep\%}
\alias{\%beep\%}
\alias{futureBeep}
\title{Beep when future is resolved}
\usage{
x \%beep\% value

futureBeep(expr, envir = parent.frame(), substitute = TRUE,
  globals = TRUE, packages = NULL, lazy = FALSE, seed = NULL,
  evaluator = plan("next"), ..., .beep_sound = 1)
}
\description{
This is pretty untested code, but the gist is that it will make a beeping sound when the future is resolved. In the meantime however, you can't change the \code{plan} and the current plan needs to have a \code{multisession} layer at the top of the hierachy, with the layer(s) you want to work on below it. This function also requires \code{beepr} to be installed. \cr
The \code{futureBeep} function is basically a wrapper of \code{future()} (and lets you specify the beep sound), and \%beep\% is basically a wrapper of \%<-\%. \cr
To-do: add docs
}
