% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_walk.R
\name{collect_decorator}
\alias{collect_decorator}
\title{Wrap a function with \code{collect_all}}
\usage{
collect_decorator(f)
}
\arguments{
\item{f}{a function}
}
\description{
This takes a function, and returns a new one, but makes it so that it goes from f(x) to collect_all(f(x)(x)). Ehhh... look at the code.
}
