% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{default_cleanup}
\alias{default_cleanup}
\alias{default_filter}
\alias{get_n_best_nodes}
\title{Default functions to help determine which cluster node to use}
\usage{
default_cleanup(df)

default_filter(df, threshold = 0.7)

get_n_best_nodes(df, n)
}
\arguments{
\item{df}{a dataframe of node information, as produced by \code{get_node_data}}

\item{threshold}{the threshold for the proportion of memory currently available by which to exclude nodes (e.g., 0.7 excludes nodes with less than 70\% of their memory currently available)}

\item{n}{the number of nodes to select}
}
\description{
These functions just apply certain default assumptions to the node information
data frames generated from \code{get_node_data}:
}
\details{
\itemize{
\item \code{default_cleanup} turns the important numeric values in the data frame (from \code{get_node_data}) from
character columns to actual numeric columns, removing the "kb" from the columns pertaining to memory,
and adding a column called \code{percent_free} which is the proportion of available memory out of the total memory.
\item \code{default_filter} simply removes nodes that aren't "free," have numbers greater than 64,
or have a proportion of free memory (\code{percent_free}) less than a certain threshold.
\item \code{get_n_best_nodes} returns \code{n} remaining nodes that have the most available memory.
}
}
