% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{default_planner}
\alias{default_planner}
\title{The default way to start a cluster plan}
\usage{
default_planner(login_node, n, ...)
}
\arguments{
\item{login_node}{a string for the gateway login, e.g., "zacharyburchill@gateway.rochester.edu" (not real)}

\item{n}{the number of clusters to use}

\item{\dots}{additional bare arguments for \code{filter()}}

\item{use_abbreviations}{because the way ssh adds known hosts, if you've only sshed into a node via \code{ssh nodeN}, you'll want to set this to \code{TRUE}}
}
\value{
By default, nothing, but possibly an expression if \code{goahead=FALSE}
}
\description{
Runs \code{future}'s plan, in a way that works well with the Rochester cluster nodes
}
