% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_n_best_nodes}
\alias{get_n_best_nodes}
\title{Returns the n best nodes}
\usage{
get_n_best_nodes(login_node, n, ..., use_abbreviations = TRUE,
  timeout_sec = 10)
}
\description{
To do: add documentation
}
