% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_node_data}
\alias{get_node_data}
\title{Convert the output of \code{pbsnodes} into a list/data frame}
\usage{
get_node_data(info_string, as_df = TRUE)
}
\arguments{
\item{info_string}{a single string of all the output of \code{pbsnodes}. Each line should be separated by a newline character}

\item{as_df}{if \code{TRUE}, it returns it as a data frame, otherwise it returns it as a list}
}
\value{
A data frame / list with all the node information
}
\description{
Runs \code{future}'s plan, in a way that works well with the Rochester cluster nodes
}
