% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{kill_r_on_nodes}
\alias{kill_r_on_nodes}
\title{Kill R processes on nodes}
\usage{
kill_r_on_nodes(node_list, secondary_login_node)
}
\description{
This command requires a "secondary" login node, i.e., a different gateway node to access the cluster nodes.  Then it just uses \code{pidof R} and \code{kill} UNIX commands to kill R. Probably only works on UNIX machines. \cr
To-do: add docs
}
