% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{monitor_resources_on_node}
\alias{monitor_resources_on_node}
\title{Monitor resource use on a single node}
\usage{
monitor_resources_on_node(username_or_command, sleeping_time = 30,
  total_checks = 10, command_maker = function(x) paste0("ps -u ", x,
  " -o pcpu,rss,size,state,time,cmd"))
}
\description{
To-do: add docs
}
