% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{\%2\%}
\alias{\%2\%}
\alias{\%3\%}
\title{Assign futures from different levels of the topology}
\usage{
x \%2\% value

x \%3\% value
}
\description{
Let's say you want to run a job on the bcs-cycle1 server, which
would require doing two nested future plans. Instead of having to
type out something like: \code{x \%<-\% { y \%<-\% { 1 + 1 }; y}}, you can
now just type \code{x \%2\% { 1 + 1 }}, which will run \code{1 + 1} on the
second layer of the future topology.  Likewise, \code{\%3\%} will assign
the expression when evaluated on the third level.
}
