% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{snap_shot_activity}
\alias{snap_shot_activity}
\title{Plot current resource consumption}
\usage{
snap_shot_activity(username_or_command, login_node, node_list, ...,
  .timeout = 30)
}
\arguments{
\item{username_or_command}{The username you're using to log in to the remote server or, if you supply \code{command_maker=NULL} to the \dots, the command you want to call and check the results of. Just stick with your username. It's easier for everyone.}

\item{login_node}{the name of the gateway node (e.g. 'zach@remote_back_up_server.server.com'). Should NOT be the same as the node you're using to run the other tasks.}

\item{node_list}{a list of the nodes you want to monitor}

\item{...}{additional arguments supplied to \code{monitor_resources_on_node}}

\item{.timeout}{The amount of seconds to wait before giving up}
}
\value{
a dataframe
}
\description{
Plots the total \code{RSS} and \code{\%CPU} from \code{\link{monitor_cluster_resources}} at the current time. \cr \cr
Basically, this is just a wrapper for \code{\link{monitor_cluster_resources}} that checks once, plots the results, and returns the data frame it used to plot them. It automatically filters out the \code{PID} of the monitoring processes.
}
