% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{test_nodes}
\alias{test_nodes}
\title{Test nodes' connectivity}
\usage{
test_nodes(node_list, login_node, timeout_sec = 2)
}
\description{
Returns nodes from node list that you can get a response from in \code{timeout_sec}. Use to filter out the bad nodes.
To-do: add docs
}
