% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_walk.R
\name{warn_about_package}
\alias{warn_about_package}
\title{Require a certain package to be installed}
\usage{
warn_about_package(package, install_location = ".")
}
\arguments{
\item{package}{a string of the package name}

\item{install_location}{a string appended to the end of the warning}
}
\description{
Eh, just look at the code. I kinda stopped using this, like a lazy, lazy coder.
}
