% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_walk.R
\name{warn_now}
\alias{warn_now}
\title{Immediately output warning conditions}
\usage{
warn_now(..., call. = TRUE, noBreaks. = FALSE, domain = NULL)
}
\arguments{
\item{\dots}{zero or more objects which can be coerced to character (and which are pasted together with no separator) or a single condition object.}

\item{call.}{logical, indicating if the call should become part of the warning message.}

\item{noBreaks.}{logical, indicating as far as possible the message should be output as a single line when \code{options(warn = 1)}.}

\item{domain}{see \code{gettext}. If NA, messages will not be translated, see also the note in \code{stop}.}
}
\description{
The function \code{base::warning} lets you immediately output warnings with
\code{immediate.=TRUE}, but not if the input is already a condition
object (weird, I know). This function is a modified Frankenstein mashup of \code{warning} and \code{message} that will immediately output warnings regardless
of their input form.
}
