% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{monitor_resources_on_node}
\alias{monitor_resources_on_node}
\title{Monitor resource use on a single node}
\usage{
monitor_resources_on_node(username_or_command, sleeping_time = 30,
  total_checks = 10, command_maker = function(x) paste0("ps -u ", x,
  " -o pcpu,rss,size,state,time,pid,cmd"), stop_file = NULL)
}
\arguments{
\item{username_or_command}{The username you're using to log in to the remote server or, if \code{command_maker} is \code{NULL}, the command you want to call and check the results of}

\item{sleeping_time}{time between checks in seconds}

\item{total_checks}{total number of checks}

\item{command_maker}{a function that takes in \code{username_or_command} or \code{NULL}}
}
\description{
To-do: add docs
}
